Move settings to stack package
