+++ internal/testrunner/runners/asset/runner.go
@@ -15,6 +15,7 @@ import (
 	"github.com/elastic/elastic-package/internal/kibana/ingestmanager"
 	"github.com/elastic/elastic-package/internal/logger"
 	"github.com/elastic/elastic-package/internal/packages"
+	"github.com/elastic/elastic-package/internal/stack"
 	"github.com/elastic/elastic-package/internal/testrunner"
 )
 
@@ -30,7 +31,7 @@ const (
 type runner struct {
 	testFolder      testrunner.TestFolder
 	packageRootPath string
-	stackSettings   testrunner.StackSettings
+	stackSettings   stack.Settings
 	esClient        *es.Client
 
 	// Execution order of following handlers is defined in runner.tearDown() method.
@@ -62,7 +63,7 @@ func (r runner) IsConfigRequired() bool {
 func (r runner) Run(options testrunner.TestOptions) ([]testrunner.TestResult, error) {
 	r.testFolder = options.TestFolder
 	r.packageRootPath = options.PackageRootPath
-	r.stackSettings = testrunner.GetStackSettingsFromEnv()
+	r.stackSettings = stack.CurrentSettings()
 	r.esClient = options.ESClient
 
 	return r.run()
